+++ tests/v2/05_images/test_filters.py
@@ -1,5 +1,5 @@
+from fractal_server.images import _filter_image_list
 from fractal_server.images import SingleImage
-from fractal_server.images.single_image import _filter_image_list
 
 images = [
     SingleImage(
